[refactor]firebase関連の処理を別フォルダに移動
